Move internal spock extension to custom package
