+++ src/integrationTest/groovy/com/wooga/spock/extensions/unity/UnityPathResolution.groovy
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2018-2020 Wooga GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.wooga.spock.extensions.unity
+
+enum UnityPathResolution {
+    /**
+     * No Unity path will be set
+     */
+    None,
+    /**
+     * A mock Unity executable will be set (a batch file that echos back)
+     */
+    Mock,
+    /**
+     * Returns an installed Unity version (if none are present, will install one locally)
+     */
+    Default
+}
