+++ src/integrationTest/groovy/com/wooga/spock/extensions/unity/UnityPluginTestOptions.groovy
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2018-2020 Wooga GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.wooga.spock.extensions.unity
+
+
+import org.spockframework.runtime.extension.ExtensionAnnotation
+import org.spockframework.runtime.extension.IAnnotationDrivenExtension
+import org.spockframework.runtime.model.FeatureInfo
+import org.spockframework.runtime.model.FieldInfo
+import org.spockframework.runtime.model.MethodInfo
+import org.spockframework.runtime.model.SpecInfo
+import wooga.gradle.unity.UnityIntegrationSpec
+
+import java.lang.annotation.Annotation
+import java.lang.annotation.ElementType
+import java.lang.annotation.Retention
+import java.lang.annotation.RetentionPolicy
+import java.lang.annotation.Target
+
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target([ElementType.METHOD, ElementType.TYPE])
+@ExtensionAnnotation(UnityPluginTestExtension.class)
+@interface UnityPluginTestOptions {
+
+    boolean applyPlugin() default true
+
+    UnityPathResolution unityPath() default UnityPathResolution.Mock
+
+    boolean addPluginTestDefaults() default true
+
+    boolean disableAutoActivateAndLicense() default true
+
+    boolean addMockTask() default true
+
+    boolean forceMockTaskRun() default true
+
+    boolean clearMockTaskActions() default false
+}
+
+
