+++ src/integrationTest/groovy/wooga/gradle/unity/UnityIntegrationSpec.groovy
@@ -17,6 +17,9 @@
 
 package wooga.gradle.unity
 
+import com.wooga.spock.extensions.unity.DefaultUnityPluginTestOptions
+import com.wooga.spock.extensions.unity.UnityPathResolution
+import com.wooga.spock.extensions.unity.UnityPluginTestOptions
 import com.wooga.spock.extensions.uvm.UnityInstallation
 import net.wooga.uvm.Installation
 import spock.lang.Shared
@@ -176,6 +179,9 @@ abstract class UnityIntegrationSpec extends IntegrationSpec {
         if (result) {
             return result
         }
+
+
+
         return preInstalledUnity.executable.path
     }
